streamlit_repo/SearchEng/requirements.txt
exa-py
streamlit
python-dotenv
